Python/requirements.txt
SpeechRecognition==3.8.1
Flask_SocketIO==3.0.2
googletrans==2.4.0
gTTS==2.0.3
opencv_python==3.4.3.18
gevent==1.3.7
apiai==1.2.3
Flask==1.0.2
pyserial==3.4
